moved models to separate module
