+++ settings.gradle.kts
@@ -1,2 +1,3 @@
 rootProject.name = "publisher_microservice"
 include("logger")
+include("model")
