+++ src/main/java/executor/service/publisher/processing/scenario/ScenarioProcessingService.java
@@ -1,6 +1,6 @@
 package executor.service.publisher.processing.scenario;
 
-import executor.service.publisher.model.Scenario;
+import executor.service.model.Scenario;
 import executor.service.publisher.processing.ProcessingService;
 
 public interface ScenarioProcessingService extends ProcessingService<Scenario> {
