+++ src/main/java/executor/service/publisher/queue/proxy/ProxySourceQueueHandler.java
@@ -1,6 +1,7 @@
 package executor.service.publisher.queue.proxy;
 
-import executor.service.publisher.model.ProxyConfigHolder;
+
+import executor.service.model.ProxyConfigHolder;
 import executor.service.publisher.queue.QueueHandler;
 
 public interface ProxySourceQueueHandler extends QueueHandler<ProxyConfigHolder> {
