+++ src/main/java/executor/service/publisher/queue/scenario/ScenarioSourceQueueHandler.java
@@ -1,6 +1,6 @@
 package executor.service.publisher.queue.scenario;
 
-import executor.service.publisher.model.Scenario;
+import executor.service.model.Scenario;
 import executor.service.publisher.queue.QueueHandler;
 
 public interface ScenarioSourceQueueHandler extends QueueHandler<Scenario> {
