+++ src/main/java/executor/service/publisher/queue/scenario/ScenarioSourceQueueHandlerImpl.java
@@ -1,7 +1,7 @@
 package executor.service.publisher.queue.scenario;
 
-import executor.service.publisher.annotation.Logged;
-import executor.service.publisher.model.Scenario;
+import executor.service.logger.annotation.Logged;
+import executor.service.model.Scenario;
 import executor.service.publisher.queue.QueueHandler;
 import lombok.RequiredArgsConstructor;
